requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
gunicorn==21.2.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
openai==0.28.1
azure-cognitiveservices-speech==1.32.1
pydub==0.25.1